+++ Adj/Program.hs
@@ -2,3 +2,4 @@ module Adj.Program (module Exports) where
 
 import Adj.Program.Instruction as Exports
 import Adj.Program.Construction as Exports
+import Adj.Program.Option as Exports